FDG-6275 moved relevant tests
